[REFACTOR] Extract duplicated code into new SymbolicMemoryClient layer
Introduce SymbolicMemoryClient(SymbolicClient) as an intermediate base for
symbolic clients that reason about tensor memory accesses. Move code that is
byte-identical between SymbolicRaceDetector and SymbolicSanitizer: the shared
launch-state attributes, _collect_tensor_base / _record_tensor_name /
_get_tensor_name / _resolve_tensor helpers, _op_make_block_ptr_overrider,
grid_idx_callback, _on_data_dependent_value, finalize, and
register_for_loop_callback. _range_to_iterator_constraint also moves up to
module scope in symbolic_engine.py.

Both subclasses rebase onto SymbolicMemoryClient. For the three MRO-shadowed
methods (grid_idx_callback, finalize, register_for_loop_callback) the
concrete subclasses add explicit forwarders calling SymbolicMemoryClient's
impl by class-qualified name — cleaner than relying on super() to thread
through the factory class.  SymbolicSanitizer._find_tensor_for_expr now
delegates to the shared _resolve_tensor before applying its nearest-segment
and first-tensor fallbacks.

No behavioral change: the 237-test suite passes unchanged.
